Updated CLI
