--- pyatlas-checks-mr/__init__.py
@@ -1,3 +1,0 @@
-import maproulette
-import os
-import json
