--- pyatlas-checks-mr/requirements.txt
@@ -1,2 +1,0 @@
-maproulette >= 1.5.0
-click
