+++ pyspark_tools/requirements.txt
@@ -1,0 +1,2 @@
+maproulette >= 1.5.0
+click
